ch4/ofi: move dynamic process code into ofi_dynproc.c
The dynamic process related code will be better maintained as separate
unit.
