+++ src/mpid/ch4/netmod/ofi/ofi_impl.h
@@ -7,6 +7,8 @@
 #define OFI_IMPL_H_INCLUDED
 
 #include <mpidimpl.h>
+#include "ofi_dynproc.h"
+/* NOTE: headers with global struct need be included before ofi_types.h */
 #include "ofi_types.h"
 #include "mpidch4r.h"
 #include "mpidig_am.h"
