+++ src/mpid/ch4/netmod/ofi/ofi_types.h
@@ -160,14 +160,6 @@ enum {
     MPIDI_OFI_PEEK_FOUND
 };
 
-enum {
-    MPIDI_OFI_DYNPROC_DISCONNECTED = 0,
-    MPIDI_OFI_DYNPROC_LOCAL_DISCONNECTED_CHILD,
-    MPIDI_OFI_DYNPROC_LOCAL_DISCONNECTED_PARENT,
-    MPIDI_OFI_DYNPROC_CONNECTED_CHILD,
-    MPIDI_OFI_DYNPROC_CONNECTED_PARENT
-};
-
 typedef struct {
     char pad[MPIDI_REQUEST_HDR_SIZE];
     struct fi_context context[MPIDI_OFI_CONTEXT_STRUCTS];       /* fixed field, do not move */
@@ -281,22 +273,6 @@ typedef struct {
     int minor_version;
     int num_am_buffers;
 } MPIDI_OFI_capabilities_t;
-
-typedef struct {
-    fi_addr_t dest;
-    int rank;
-    int state;
-} MPIDI_OFI_conn_t;
-
-typedef struct MPIDI_OFI_conn_manager_t {
-    int max_n_conn;             /* Maximum number of connections up to this point */
-    int n_conn;                 /* Current number of open connections */
-    int next_conn_id;           /* The next connection id to be used. */
-    int *free_conn_id;          /* The list of the next connection id to be used so we
-                                 * can garbage collect as we go. */
-    MPIDI_OFI_conn_t *conn_list;        /* The list of connection structs to track the
-                                         * outstanding dynamic process connections. */
-} MPIDI_OFI_conn_manager_t;
 
 /* Global state data */
 #define MPIDI_KVSAPPSTRLEN 1024
